UHF-11015: Move getAttachments to Decision bundle class
